Try to get tests working a bit more, they seem to at least run now
